Create windowCleaner.py
Python code developed for a Raspberry Pi controller. The code features the use of timers, switches, an accelerometer, and a gyroscope, as well as the implementation of a path planning algorithm. The window cleaner attaches to the surface of a window and cleans it by traversing across the window in a procession of horizontal segments. 
Please visit the project website for more details: https://sites.google.com/site/cmumechatronics2015teami2/